Major code refactor to implement interactable blocks
Removed a lot of responsibilities of the level class as it was starting to do too much, I felt that a level wasn't really responsible for dealing with entity and tilecollision and drawing the scene.  This was brought about by wanting interactable block (e.g. when Mario bumps into one from below), it was simpler to just have the tiles in the level class referenced by functions to deal with the functionality rather than interacting with different classes.  Packaged collision functionality into the corresponding trait classes (which makes the Koopa behaviour simpler) and removed TileCollider and EntityCollider classes as they seemed pointless.