Can now bump blocks from underneath and they briefly nudge up and back down
Had to store the x and y positions of each tile in the tile grid stored in the level class and draw to the screen using this rather than the index of each tile.  This may mean that the use of indexing tiles could go, will have to think about it.