Minimised number of draw calls in drawing background layer
Noticed frame rate wavering and all the time seems to be in requestAnimationFrame so just trying to minimise the amount of drawing done.  A lot of advice online suggests generating an offscreen canvas for generating the image in each frame and then drawing this to your main canvas but it seems to be slower, need to keep an eye on this as I intend on making another draw layer.